Deduplication of code between FhirBundleGenerator and FhirGenerationPipeline. Bound resource dates to report start/end dates. Code Rabbit Feedback
