+++ Tests/BackendE2ETests/MultiMeasureTest.cs
@@ -127,8 +127,8 @@ public sealed class MultiMeasureTest : IAsyncLifetime, IClassFixture<BackendE2ET
             acquisitionSimulation: new FhirGenerationPipeline.AcquisitionSimulationConfig
             {
                 QueryPlan = QueryPlanBuilder.GetDefaultAsInput(),
-                ReportStart = Config.StartDate,
-                ReportEnd = Config.EndDate
+                ClinicalPeriodStart = Config.StartDate,
+                ClinicalPeriodEnd = Config.EndDate
             });
 
         _generationManifest = pipelineResult.Manifest;
