+++ Tests/BackendE2ETests/ReportScheduledWorkflowTest.cs
@@ -75,8 +75,8 @@ public sealed class ReportScheduledTest : IAsyncLifetime, IClassFixture<BackendE
             acquisitionSimulation: new FhirGenerationPipeline.AcquisitionSimulationConfig
             {
                 QueryPlan = QueryPlanBuilder.GetDefaultAsInput(),
-                ReportStart = _config.StartDate,
-                ReportEnd = _config.EndDate
+                ClinicalPeriodStart = _config.StartDate,
+                ClinicalPeriodEnd = _config.EndDate
             });
 
         _generationManifest = pipelineResult.Manifest;
